renaming the SelectWithAutomCompleteComponent to ComboboxComponent
