+++ app/components/combobox_component.html.erb
@@ -1,0 +1,30 @@
+<div
+  data-controller="combobox"
+  data-combobox-trigger-type-value="click"
+  data-combobox-strategy-value="fixed"
+  data-combobox-no-results-text-value="<%= t(".no_results_text") %>"
+  data-combobox-single-result-text-value="<%= t(".single_result_text") %>"
+  data-combobox-multiple-results-text-value="<%= t(".multiple_results_text") %>"
+>
+  <%= @form.hidden_field @field,
+                     value: @selected_option[:value],
+                     id: "#{@trigger_id}_hidden",
+                     data: {
+                       "combobox-target": "hidden",
+                     } %>
+  <%= render Viral::BaseComponent.new(**trigger_arguments) %>
+  <div
+    id="<%= @menu_id %>"
+    role="listbox"
+    aria-label="<%= trigger_arguments[:aria][:label] %>"
+    data-combobox-target="menu"
+    class="
+      fixed z-10 m-0 box-border hidden max-h-[250px] overflow-x-hidden overflow-y-auto
+      border border-slate-500 bg-white p-0 dark:border-slate-600 dark:bg-slate-800
+    "
+  >
+    <%= @menu_options %>
+  </div>
+  <div class="sr-only" aria-live="polite" data-combobox-target="ariaLiveUpdate">
+  </div>
+</div>
