--- app/components/select_with_auto_complete_component.html.erb
@@ -1,34 +1,0 @@
-<div
-  data-controller="select-with-auto-complete"
-  data-select-with-auto-complete-trigger-type-value="click"
-  data-select-with-auto-complete-strategy-value="fixed"
-  data-select-with-auto-complete-no-results-text-value="<%= t(".no_results_text") %>"
-  data-select-with-auto-complete-single-result-text-value="<%= t(".single_result_text") %>"
-  data-select-with-auto-complete-multiple-results-text-value="<%= t(".multiple_results_text") %>"
->
-  <%= @form.hidden_field @field,
-                     value: @selected_option[:value],
-                     id: "#{@trigger_id}_hidden",
-                     data: {
-                       "select-with-auto-complete-target": "hidden",
-                     } %>
-  <%= render Viral::BaseComponent.new(**trigger_arguments) %>
-  <div
-    id="<%= @menu_id %>"
-    role="listbox"
-    aria-label="<%= trigger_arguments[:aria][:label] %>"
-    data-select-with-auto-complete-target="menu"
-    class="
-      fixed z-10 m-0 box-border hidden max-h-[250px] overflow-x-hidden overflow-y-auto
-      border border-slate-500 bg-white p-0 dark:border-slate-600 dark:bg-slate-800
-    "
-  >
-    <%= @menu_options %>
-  </div>
-  <div
-    class="sr-only"
-    aria-live="polite"
-    data-select-with-auto-complete-target="ariaLiveUpdate"
-  >
-  </div>
-</div>
