+++ config/locales/en.yml
@@ -331,6 +331,10 @@ en:
           expiration: Expiration
           username: Username
         never: Never
+  combobox_component:
+    multiple_results_text: "%{num} results found"
+    no_results_text: No results found
+    single_result_text: 1 result found
   concerns:
     attachment_actions:
       destroy:
@@ -1710,10 +1714,6 @@ en:
         select_project: Select project
         submit_button: Submit
         title: Transfer Samples
-  select_with_auto_complete_component:
-    multiple_results_text: "%{num} results found"
-    no_results_text: No results found
-    single_result_text: 1 result found
   services:
     attachments:
       concatenation:
