+++ config/locales/fr.yml
@@ -331,6 +331,10 @@ fr:
           expiration: Expiration
           username: Nom d’utilisateur
         never: Jamais
+  combobox_component:
+    multiple_results_text: "%{num} résultats trouvés"
+    no_results_text: Aucun résultat trouvé
+    single_result_text: 1 résultat trouvé
   concerns:
     attachment_actions:
       destroy:
@@ -1714,10 +1718,6 @@ fr:
         select_project: Sélectionner le projet
         submit_button: Envoyer
         title: Transférer des échantillons
-  select_with_auto_complete_component:
-    multiple_results_text: "%{num} résultats trouvés"
-    no_results_text: Aucun résultat trouvé
-    single_result_text: 1 résultat trouvé
   services:
     attachments:
       concatenation:
