+++ test/components/previews/combobox_component_preview/default.html.erb
@@ -1,0 +1,4 @@
+<%= form_with(url: "#") do |form| %>
+  <label for="<%=form.field_id(:field)%>">Metadata field</label>
+  <%= render ComboboxComponent.new(form: form, field: :field, options: options) %>
+<% end %>
