--- test/components/previews/select_with_auto_complete_component_preview/default.html.erb
@@ -1,8 +1,0 @@
-<%= form_with(url: "#") do |form| %>
-  <label for="<%=form.field_id(:field)%>">Metadata field</label>
-  <%= render SelectWithAutoCompleteComponent.new(
-    form: form,
-    field: :field,
-    options: options,
-  ) %>
-<% end %>
